+++ CHANGELOG.md
@@ -74,6 +74,7 @@ Note that since we don't clearly distinguish between a public and private interf
     - Add `AnimateStateSnapshotTransition` animation
     - Update the snapshots UI to support this feature
 - Use "proper time" in the animation loop to prevent animation skips during blocking operations (e.g., shader complication)
+- Add `Color.interpolateHcl`
 
 ## [v4.18.0] - 2025-06-08
 - MolViewSpec extension: